backtesting_algo/requirements.txt
python-binance
TA-Lib
numpy
backtrader
